+++ lib/africa_schools.dart
@@ -1,7 +1,4 @@
 library africa_schools;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/nigeria.dart';
+