+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.1
+# Changelog for Africa Schools
 
-* TODO: Describe initial release.
+A package that makes you get access to African Tertiary Institutions.
+
+## 1.0.0
+### List of Strings for Nigerian Institutions
+- Nigeria's Universities added
+- Nigeria's Polytechnics added
+- Nigeria's Colleges of Education added
+- Nigeria's Health Science added